why is this happening on windows

+++ cli/htmgo/tasks/process/pid_unix.go
@@ -10,6 +10,9 @@ import (
 )
 
 func KillProcess(process *os.Process) error {
+	if process == nil {
+		return nil
+	}
 	return syscall.Kill(-process.Pid, syscall.SIGKILL)
 }
